refactor: 절대 경로로 모듈 임포트 정리 및 interrogation 이름 충돌 해결
